requirements.txt
absl-py==2.1.0
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
comm==0.2.2
confection==0.1.5
cymem==2.0.8
datasets==2.20.0
debugpy==1.8.2
decorator==5.1.1
dill==0.3.8
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
evaluate==0.4.2
executing==2.0.1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
gensim==4.3.2
google-auth==2.31.0
google-auth-oauthlib==1.0.0
grpcio==1.64.1
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.0.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
langcodes==3.4.0
language_data==1.2.0
lightning-utilities==0.11.3.post0
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
numpy==1.24.4
oauthlib==3.2.2
packaging==24.1
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
platformdirs==4.2.2
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
pyDeprecate==0.3.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytorch-lightning==1.5.8
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rouge_score==0.1.2
rsa==4.9
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.12.1
tbb==2021.13.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
torchmetrics==1.4.0.post0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers @ file:///D:/graduatiion_projict/transformers
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
